Fix Redundant Declerations
Remove some redundant declerations and resolve a circular header dependency
by splitting out the ssl types.
