+++ library/ssl_debug_helpers.h
@@ -11,11 +11,10 @@
 #ifndef MBEDTLS_SSL_DEBUG_HELPERS_H
 #define MBEDTLS_SSL_DEBUG_HELPERS_H
 
-#include "ssl_misc.h"
-
 #if defined(MBEDTLS_DEBUG_C)
 
 #include "mbedtls/ssl.h"
+#include "ssl_types.h"
 
 const char *mbedtls_ssl_states_str(mbedtls_ssl_states in);
